move convert_labels_to_integers from utils/utilities.py to classes/function.py. The reason is that it was causing aboslute/relative import issues
