+++ xgi/classes/function.py
@@ -3,6 +3,7 @@
 from collections import Counter
 
 from ..exception import XGIError
+from .hypergraph import Hypergraph
 
 __all__ = [
     "max_edge_order",
@@ -21,6 +22,7 @@ __all__ = [
     "get_edge_attributes",
     "is_empty",
     "maximal_simplices",
+    "convert_labels_to_integers",
 ]
 
 
@@ -613,3 +615,44 @@ def maximal_simplices(SC):
         if maximal:
             max_simplices.append(i)
     return max_simplices
+
+
+def convert_labels_to_integers(H, label_attribute="label"):
+    """Relabel node and edge IDs to be sequential integers.
+
+    Parameters
+    ----------
+    H : Hypergraph
+        The hypergraph of interest
+
+    label_attribute : string, default: "label"
+        The attribute name that stores the old node and edge labels
+
+    Returns
+    -------
+    Hypergraph
+        A new hypergraph with nodes and edges with sequential IDs starting at 0.
+        The old IDs are stored in the "label" attribute for both nodes and edges.
+
+    Notes
+    -----
+    The "relabeling" will occur even if the node/edge IDs are sequential.
+    Because the old IDs are stored in the "label" attribute for both nodes and edges,
+    the old "label" values (if they exist) will be overwritten.
+    """
+    node_dict = dict(zip(H.nodes, range(H.num_nodes)))
+    edge_dict = dict(zip(H.edges, range(H.num_edges)))
+    temp_H = Hypergraph()
+    temp_H._hypergraph = H._hypergraph.copy()
+
+    for node, id in node_dict.items():
+        temp_H._node[id] = [edge_dict[e] for e in H._node[node]]
+        temp_H._node_attr[id] = H._node_attr[node].copy()
+        temp_H._node_attr[id][label_attribute] = node
+
+    for edge, id in edge_dict.items():
+        temp_H._edge[id] = [node_dict[n] for n in H._edge[edge]]
+        temp_H._edge_attr[id] = H._edge_attr[edge].copy()
+        temp_H._edge_attr[id][label_attribute] = edge
+
+    return temp_H
