+++ xgi/utils/utilities.py
@@ -1,9 +1,8 @@
 """General utilities."""
-from collections import defaultdict
 
-from ..classes import Hypergraph
+from collections import defaultdict, Counter
 
-__all__ = ["dual_dict", "convert_labels_to_integers"]
+__all__ = ["dual_dict"]
 
 
 def dual_dict(edge_dict):
@@ -36,44 +35,3 @@ def dual_dict(edge_dict):
             node_dict[node].append(edge_id)
 
     return dict(node_dict)
-
-
-def convert_labels_to_integers(H, label_attribute="label"):
-    """Relabel node and edge IDs to be sequential integers.
-
-    Parameters
-    ----------
-    H : Hypergraph
-        The hypergraph of interest
-
-    label_attribute : string, default: "label"
-        The attribute name that stores the old node and edge labels
-
-    Returns
-    -------
-    Hypergraph
-        A new hypergraph with nodes and edges with sequential IDs starting at 0.
-        The old IDs are stored in the "label" attribute for both nodes and edges.
-
-    Notes
-    -----
-    The "relabeling" will occur even if the node/edge IDs are sequential.
-    Because the old IDs are stored in the "label" attribute for both nodes and edges,
-    the old "label" values (if they exist) will be overwritten.
-    """
-    node_dict = dict(zip(H.nodes, range(H.num_nodes)))
-    edge_dict = dict(zip(H.edges, range(H.num_edges)))
-    temp_H = Hypergraph()
-    temp_H._hypergraph = H._hypergraph.copy()
-
-    for node, id in node_dict.items():
-        temp_H._node[id] = [edge_dict[e] for e in H._node[node]]
-        temp_H._node_attr[id] = H._node_attr[node].copy()
-        temp_H._node_attr[id][label_attribute] = node
-
-    for edge, id in edge_dict.items():
-        temp_H._edge[id] = [node_dict[n] for n in H._edge[edge]]
-        temp_H._edge_attr[id] = H._edge_attr[edge].copy()
-        temp_H._edge_attr[id][label_attribute] = edge
-
-    return temp_H
